Change schemas and column types/seed db

+++ db/migrate/20171126030523_create_beer_reviews.rb
@@ -1,7 +1,7 @@
 class CreateBeerReviews < ActiveRecord::Migration[5.1]
   def change
     create_table :beer_reviews do |t|
-      t.integer :rating, null: false
+      t.float :rating, null: false
       t.text :review
       t.integer :user_id, null: false
       t.integer :beer_id, null: false
